refactor: next함수 리팩토링
