+++ src/app/channel/_components/player/index.tsx
@@ -2,7 +2,7 @@
 
 import { ChangeEvent, useEffect, useLayoutEffect, useRef, useState } from 'react';
 
-import send from '@/services/websocket/send';
+import { nextVideo, send } from '@/services/websocket';
 import { VideoMoveReq, VideoPlayReq } from '@/services/websocket/type';
 import { useUserStore } from '@/stores/User';
 import { useSocketStore } from '@/stores/useSocketStore';
@@ -10,7 +10,6 @@ import ReactPlayer from 'react-player';
 import { useShallow } from 'zustand/react/shallow';
 
 import './player.css';
-import usePlaylist from '../../_hooks/usePlaylist';
 
 const YOUTUBE_BASE_URL = 'https://www.youtube.com/watch?v=';
 
@@ -20,7 +19,6 @@ interface Props {
 }
 
 const Player = ({ channelId, owner_id }: Props) => {
-  const { next } = usePlaylist();
   const { video, setStore } = useSocketStore(
     useShallow((state) => ({ video: state.video, setStore: state.setSocketStore })),
   );
@@ -60,7 +58,7 @@ const Player = ({ channelId, owner_id }: Props) => {
   };
 
   const onEnded = () => {
-    next();
+    nextVideo();
   };
 
   const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
