+++ src/app/channel/_components/playlist/index.tsx
@@ -8,7 +8,7 @@ import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { GetChannelDetailResponse } from '@/services/channel/type';
 import { getPlaylist } from '@/services/playlist';
-import send from '@/services/websocket/send';
+import { nextVideo, send } from '@/services/websocket';
 import { useUserStore } from '@/stores/User';
 import { useSocketStore } from '@/stores/useSocketStore';
 import { ListVideo, Plus, Trash2 } from 'lucide-react';
@@ -52,10 +52,7 @@ const Playlist = ({ channel, owner }: Props) => {
   /** 비디오 재생 */
   const handlePlay = (index: number) => {
     if (!isOwner) return;
-    send('VIDEO_PLAY', {
-      playlist_no: playlist[index].playlist_no,
-      playing: true,
-    });
+    nextVideo(index);
   };
 
   /** 비디오 삭제 */
