+++ src/services/websocket/index.ts
@@ -1,50 +1,5 @@
-import { Client } from '@stomp/stompjs';
-import SockJS from 'sockjs-client';
+import nextVideo from './nextVideo';
+import send from './send';
+import { stomp } from './stompClient';
 
-import { useSocketStore } from '@/stores/useSocketStore';
-import { getSession } from 'next-auth/react';
-
-import type { IFrame, IMessage, IStompSocket } from '@stomp/stompjs';
-
-const createStompClient = () => {
-  const client = new Client();
-  const BASE_URL = process.env.NEXT_PUBLIC_WEBSOCKET_SERVER_URL;
-
-  const { isConnected, setIsConnected } = useSocketStore.getState();
-
-  /** 소켓 연결 */
-  const connect = async (channelId: string, onMessageCallback: (message: IMessage) => void) => {
-    const session = await getSession();
-
-    const url = await new URL(
-      `/ws?channel=${channelId}${session?.user.accessToken ? `&token=${session.user.accessToken}` : ''}`,
-      BASE_URL,
-    ).toString();
-    client.webSocketFactory = () => new SockJS(url) as IStompSocket;
-
-    client.onConnect = () => {
-      client.subscribe(`/channel/${channelId}`, onMessageCallback);
-      setIsConnected(true);
-    };
-
-    client.onStompError = (frame: IFrame) => {
-      console.error(frame.headers['message']);
-      console.error(frame.body);
-    };
-
-    client.activate();
-  };
-
-  /** 소캣 연결 해제 */
-  const disconnect = () => {
-    client.deactivate();
-    setIsConnected(false);
-  };
-
-  const send = (endpoint: string, message: object) => {
-    client.publish({ destination: endpoint, body: JSON.stringify(message) });
-  };
-  return { connect, disconnect, send, isConnected };
-};
-
-export const stomp = createStompClient();
+export { send, stomp, nextVideo };
