+++ src/services/websocket/nextVideo.ts
@@ -1,0 +1,25 @@
+import { useSocketStore } from '@/stores/useSocketStore';
+
+import send from './send';
+
+const nextVideo = (index?: number) => {
+  const { playlist, video } = useSocketStore.getState();
+
+  let nextIndex;
+
+  if (index) {
+    nextIndex = index;
+  } else {
+    nextIndex = playlist.findIndex((item) => item.playlist_no === video.playlist_no) + 1;
+    if (playlist.length <= nextIndex) {
+      nextIndex = 0;
+    }
+  }
+
+  const { playlist_no } = playlist[nextIndex];
+
+  send('VIDEO_PLAY', { playlist_no, playing: true });
+  send('VIDEO_MOVE', { playlist_no, playtime: 0 });
+};
+
+export default nextVideo;
